Add Airplane, refactoring.

+++ Aircrafts/Interfaces/IFlyable.cs
@@ -6,6 +6,15 @@ namespace Aircrafts.Interfaces
     {
         public void FlyTO(Point3D point);
 
-        public uint GetFlyTime(Point3D point);
+        /// <summary>
+        /// Calculate flying time between two points.
+        /// </summary>
+        /// <param name="point">
+        /// The point to which the flight time is calculated
+        /// </param>
+        /// <returns>
+        /// Returns the flying time to the point in double format.
+        /// </returns>
+        public double GetFlyTime(Point3D point);
     }
 }
